prevent a gap appearing behind the keyboard when it goes away and refactor the code a little
